refactor work queues: add schedule test for workers > 1
We can use the finally callback to know when a tag is available for scheduling
again.
Although we cannot check that the resulting execution order is completely valid,
(that would be very timing dependent due to the combination of RR and per-tag
 serialisation), we can check some weaker properties:#
 * operations are executed in order per-tag
 * the only reason to schedule the same tag again is if all the tags that are
schedulable were already seen during the current RR cycle,
where schedulable is computed based on the currently executing tags,
and a RR cycle starts when all the tags that can be scheduled were scheduled at
least once.
(This essentially checks that operations from a tag do not starve others)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>