CP-28671: use -device -drive instead of just drive in QEMU
In preparation for adding -bootindex, which can only be added on the device, and not the drive.

`index=` also has no effect on `-drive` when `if=none` and a `-device` is used.
Instead `bus` and `unit` has to be used. For IDE there are two buses:
`ide.0` and `ide.1` each with 2 units (QEMU correctly refuses to accept more than 2).

The device name should also not be `cd` when the device is a disk.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>