CP-37034: move TSX handling logic out of xenopsd
There have been 2 microcode updates on Intel that changed the behaviour
of the TSX instruction.
Currently there is logic in xenopsd to query MSR_ARCH_CAPS, and look at
TSX_CTRL and enable HLE/RTM in the migration-safety CPUID compatibility
featureset.
However the 2nd microcode update moved the location of that information,
and it is not always present in that MSR anymore.
Xen currently "fakes up" the TSX_CTRL bit for backwards compatibility
reasons, but this just adds technical debt.

Reuse the already existing logic in Xen instead of doing this calculation
in the toolstack which becomes outdated on microcode and Xen updates.

Query the HVM/PV Max policy from Xen and logically OR those feature bits
into `features_hvm` and `features_pv` respectively.

On latest upstream Xen the Max policy tells us what features we can
accept on an incoming migration: i.e. the features are implemented
natively in hardware, emulated by Xen or a compatible workaround by
hardware.
As an example: the TSX instruction may be hidden in the default policy,
i.e. we do not want guests to see it and use it on a fresh boot,
but present in the Max policy.
This is safe because when the host has a TSX instruction the microcode
update would just make the TSX instruction always abort, which is
compliant with the instruction's specification (but then it is not a
very useful instruction anymore, and better hidden on boot).

On Xen 4.13.x there is no Max/default policy split yet, but xenctrl
in the patchqueue can make the necessary adjustments for TSX in the
short term while we work on updating to a newer version of Xen on
Stockholm.

This requires a version of Xen that has the new xenctrl stubs.

Drops a lot of technical debt and prints various information about
policies for debugging purposes (including diffing policies to make it
obvious what the difference is between them).

The featureset can be decoded into human readable form by
pasting it as argument to the `xen-cpuid` executable.
The policy can be decoded or created by looking up the values in the
Intel/AMD CPU manuals (see xenctrl.mli for details).

The policy is not currently stored, but when we get CPUID/MSR leveling
we will store in in XAPI. Thus the string serialization/deserialization
code is here in xenctrlext.ml instead of xenctrl.ml so that the format
is under the toolstack's control.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>